Reorganize query tree unit tests
* Breaking source into multiple files
* Enable control of the instantiation for tree and device type
